Change position type to decimal.

+++ server/boards/models.py
@@ -20,7 +20,7 @@ class List(models.Model):
     title = models.CharField(max_length=50)
     board = models.ForeignKey(Board, on_delete=models.CASCADE)
     is_archived = models.BooleanField(default=False)
-    position = models.PositiveIntegerField(unique=True)
+    position = models.DecimalField(unique=True, max_digits=5, decimal_places=5)
 
     def __str__(self):
         return self.title
